// src/modules/invoice/usecase/find-all-invoice/find-all-invoice.usecase.dto.ts
import Address from '../../../@shared/domain/value-object/address';
import InvoiceItems from '../../domain/invoice.itens.entity';

export interface FindAllInvoiceUseCaseOutputDTO {
  id: string;
  name: string;
  document: string;
  address: Address;
  items: InvoiceItems[];
  createdAt: Date;
  updatedAt: Date;
}